backend/requirements.txt
Django==1.11.2
django-cors-headers==2.1.0
django-extensions==1.7.9
djangorestframework==3.6.3
psycopg2==2.7.1
pytesseract==0.1.7
